// src/types.ts
export enum ViewState {
  DASHBOARD_WINF = 'WINF',
  DASHBOARD_ASCEND = 'ASCEND',
  DASHBOARD_SYNAPSE = 'SYNAPSE',
  W_RANK = 'W_RANK',
  DASHBOARD_WINFCOIN = 'WINFCOIN',
  SALES_FUNNEL = 'FUNNEL',
  WARRANTY = 'WARRANTY',
  MODULES = 'MODULES',
  PROFILE = 'PROFILE',
  MODULE_ECOTECH = 'MODULE_ECOTECH',
  MODULE_FACECAR = 'MODULE_FACECAR',
  MODULE_GRID = 'MODULE_GRID',
  MODULE_CROSS_PLATFORM = 'MODULE_CROSS_PLATFORM',
  MODULE_NEUROMESH = 'MODULE_NEUROMESH',
  MODULE_CAPTURE = 'MODULE_CAPTURE',
  MODULE_MEMBERS = 'MODULE_MEMBERS',
  MODULE_SYNTHWAVE = 'MODULE_SYNTHWAVE',
  MODULE_UNIVERSO_DARK = 'MODULE_UNIVERSO_DARK',
  LANDING_UNIVERSO_DARK = 'LANDING_UNIVERSO_DARK',
  MODULE_WARP = 'MODULE_WARP',
  MODULE_CONCIERGE = 'MODULE_CONCIERGE',
  MODULE_DATA_CORE = 'MODULE_DATA_CORE',
  MODULE_ACADEMY = 'ACADEMY',
  MODULE_ESCAPE_3D = 'MODULE_ESCAPE_3D',
  MODULE_PRICE_TABLE = 'MODULE_PRICE_TABLE',
  MODULE_BIOMETRIC = 'MODULE_BIOMETRIC',
  MODULE_FINANCIAL = 'MODULE_FINANCIAL',
  MODULE_QUOTES = 'MODULE_QUOTES',
  MODULE_NEURO_PARADOX = 'NEURO_PARADOX',
  MODULE_NEURAL_FLOW = 'NEURAL_FLOW',
  MODULE_SQUAD_PERFORMANCE = 'SQUAD_PERFORMANCE',
  MODULE_ARCHITECTURAL = 'ARCHITECTURAL',
  MODULE_ARSENAL = 'ARSENAL',
  MODULE_KIOSK_MODE = 'KIOSK_MODE',
  BRAND_MANUAL = 'BRAND_MANUAL',
  MODULE_WINF_CUT = 'WINF_CUT',
  MODULE_AVATAR_JOURNEY = 'AVATAR_JOURNEY',
  MODULE_STOCK = 'STOCK',
  MODULE_THE_BOARD = 'THE_BOARD',
  MODULE_MISSION_CONTROL = 'MISSION_CONTROL',
  MODULE_CONNECT = 'CONNECT',
  MODULE_NEURAL_BRIDGE = 'NEURAL_BRIDGE',
  MODULE_WHATSAPP_HUB = 'WHATSAPP_HUB',
  MODULE_CONSULTANCY_LINK = 'CONSULTANCY_LINK',
  MODULE_TASKS = 'MODULE_TASKS',
  MODULE_FAQ = 'MODULE_FAQ',
  MODULE_PRODUCTS = 'MODULE_PRODUCTS',
  MODULE_INSTALLATIONS = 'MODULE_INSTALLATIONS',
  MODULE_BLACKSHOP_ADMIN = 'MODULE_BLACKSHOP_ADMIN',
  MODULE_WARRANTIES = 'MODULE_WARRANTIES',
  CERTIFICATE_VIEWER = 'CERTIFICATE_VIEWER',
  PRODUCTS_CATALOG = 'PRODUCTS_CATALOG',
  PRODUCT_LANDING = 'PRODUCT_LANDING',
  MODULE_BLACKSHOP = 'MODULE_BLACKSHOP',
  MODULE_CHECKOUT_BLACKSHOP = 'MODULE_CHECKOUT_BLACKSHOP',
  MODULE_WINF_CHAIN = 'MODULE_WINF_CHAIN',
  MODULE_CORE_AI = 'MODULE_CORE_AI',
  MODULE_NEXUS_AI = 'MODULE_NEXUS_AI',
  MODULE_WINF_BRAIN = 'MODULE_WINF_BRAIN',
  MODULE_WINF_WORLD = 'MODULE_WINF_WORLD',
  MODULE_INTEGRATIONS = 'MODULE_INTEGRATIONS',
  MODULE_MARKETING_AI = 'MODULE_MARKETING_AI',
  MODULE_CONTROL_ROOM = 'MODULE_CONTROL_ROOM',
  PUBLIC_PORTAL = 'PUBLIC_PORTAL',
  DASHBOARD_ARCHITECT = 'DASHBOARD_ARCHITECT',
  DASHBOARD_ASSET_LIGHT = 'DASHBOARD_ASSET_LIGHT',
  MODULE_ESSENTIAL_TOOLS = 'ESSENTIAL_TOOLS',
  
  LANDING_PAGE = 'LANDING_PAGE',
  LANDING_PAGE_MENU = 'LANDING_PAGE_MENU',
  INSTITUTIONAL_SITE = 'INSTITUTIONAL_SITE',
  PUBLIC_CONSULTANCY = 'PUBLIC_CONSULTANCY',
  PUBLIC_CONSULTANCY_ENTRANCE = 'PUBLIC_CONSULTANCY_ENTRANCE',
  LANDING_PARCERIA = 'LANDING_PARCERIA',
  LANDING_ASSET_LIGHT = 'LANDING_ASSET_LIGHT',
  LANDING_PARCERIA_NIVEL1 = 'LANDING_PARCERIA_NIVEL1',
  LANDING_PARCERIA_NIVEL2 = 'LANDING_PARCERIA_NIVEL2',
  LANDING_PARCERIA_NIVEL3 = 'LANDING_PARCERIA_NIVEL3',
  LANDING_KIOSK = 'LANDING_KIOSK',
  LANDING_STUDIO = 'LANDING_STUDIO',
  LANDING_AEROCORE = 'LANDING_AEROCORE',
  LANDING_NEOSKIN = 'LANDING_NEOSKIN',
  LANDING_BIOMETRIC = 'LANDING_BIOMETRIC',
  LANDING_WINF_SELECT_ELITE = 'LANDING_WINF_SELECT_ELITE',
  ABOUT_US = 'ABOUT_US',
  ACCESS_FORM = 'ACCESS_FORM',
  ARCHITECT_REGISTRATION = 'ARCHITECT_REGISTRATION',
  LANDING_PARTNERS_CORPORATE = 'LANDING_PARTNERS_CORPORATE',
  LANDING_GALLERY = 'LANDING_GALLERY',
  
  LOGIN = 'LOGIN',
  SEARCH = 'SEARCH',
  ADMIN_PANEL = 'ADMIN_PANEL',
  MODULE_MOLECULAR_TWIN = 'MOLECULAR_TWIN',
  MODULE_CUIDAR_CONECTAR = 'CUIDAR_CONECTAR',
  MODULE_DIGITAL_START = 'DIGITAL_START',
  MODULE_SHOP_CORE = 'SHOP_CORE',
  WINF_BRAND_LANDING = 'WINF_BRAND_LANDING',
  LANDING_SYNTHWAVE = 'LANDING_SYNTHWAVE',
  LANDING_BRIEFING = 'LANDING_BRIEFING',
  BRAND_SHOWROOM = 'BRAND_SHOWROOM',
}